Update dependencies and readme
Be a bit more lenient with dependencies' versions.
As of python-lzo 1.15, wheels are provided for all recent versions of
CPython on Windows (AMD64 only).
As of lxml 4.9.2, wheels are provided for CPython 3.11 on Windows and
CPython 3.10 on macOS.
ubi_reader 0.8.7 requires lzallright >= 0.1.0 which has
vlaci/lzallright#12 and 0.8.8 requires lzallright >= 0.2.1 which has
vlaci/lzallright#20.